Add files via upload
Number Sequence 2 🌶️
Given two integers
   
m and
   
n. Write a program that prints all numbers from
   
m to
   
n inclusive in ascending order if
   
<
   
m<n, or in descending order otherwise.

Input data format
The program takes two integers as input.
   
m and
   
n, each on a separate line.

Output format
The program should display the numbers in accordance with the condition of the problem.