Add files via upload
Amount of numbers
The program takes two integers as input.
   
a and
   
b
(
   
≤
   
)
(a≤b). Write a program that counts the number of numbers in the range from
   
a to
   
b inclusive, whose cube ends in
4
4 or
9
9.

Input data format
The program takes two integers as input.
   
a and
   
b
(
   
≤
   
)
(a≤b).

Output format
The program should output one integer according to the program condition.

Note. cube number
   
a is its third power
   
3
a
3
 .

Test Data   
Sample Input 1:

1
10
Sample Output 1:

2