Add files via upload
Correct email
We will consider an email address to be valid if it contains the dog symbol (@) and dots. Write a program that checks the correctness of an email address.

Input data format
The input to the program is one string - an email address.

Output format
The program should output the string "YES" if the email address is valid and "NO" otherwise.

Note. The presence of @ and . is not enough for the correctness of the email address, but their absence is guaranteed to entail an incorrect email address.

Test Data 🟢
Sample Input 1:

aaaa@bbb.com
Sample Output 1:

YES
Sample Input 2:

aaaa@bbb.com
Sample Output 2:

NO