Add files via upload
Factorial
The input to the program is a natural number
�
n. Write a program that calculates
�
!
n!.

Input data
The input to the program is a natural number
�
,
(
�
≤
12
)
n,(n≤12).

Output
The program should display a single number in accordance with the condition of the problem.

Note. Factorial of a natural number
�
n, is called the product of all natural numbers from
1
1 to
�
n, that is
�
!
=
1
⋅
2
⋅
3
⋅
…
⋅
�
n!=1⋅2⋅3⋅…⋅n

Test Data 🟢
Sample Input 1:

3
Sample Output 1:

6
Sample Input 2:

1
Sample Output 2:

1
Sample Input 3:

2
Sample Output 3:

2