Add files via upload
Sum of numbers 2
The input to the program is a natural number
�
n. Write a program that calculates the sum of those numbers from
1
1 to
�
n (inclusive) whose square ends in
2
,
5
2.5 or
8
8.

Input data format
The input to the program is a natural number
�
n.

Output format
The program should display a single number in accordance with the condition of the problem.

Note. If there are no such numbers in the specified range, then output
0
0.

Test Data 🟢
Sample Input 1:

10
Sample Output 1:

5
Sample Input 2:

100
Sample Output 2:

500