Add files via upload
Until the END 2
The input to the program is a sequence of words, each word on a separate line. The end of the sequence is the word "END" or "end" (capital or small letters, no quotes). At the same time, the words “END” and “end” themselves are not included in the sequence, only symbolizing its end. Write a program that prints the terms of a given sequence.

Input data format
The input to the program is a sequence of words, each word on a separate line.

Output format
The program should output the terms of the given sequence.

Test Data 🟢
Sample Input 1:

JavaScript
C++
C#
ruby
PHP
END
Python
Sample Output 1:

JavaScript
C++
C#
ruby
PHP